+++ common/src/main/java/org/bahmni/form2/common/db/JDBCConnectionProvider.java
@@ -1,0 +1,9 @@
+package org.bahmni.form2.common.db;
+
+import java.sql.Connection;
+
+public interface JDBCConnectionProvider {
+    public Connection getConnection();
+
+    public void closeConnection();
+}
