--- common/src/main/java/org/bahmni/module/common/db/JDBCConnectionProvider.java
@@ -1,9 +1,0 @@
-package org.bahmni.module.common.db;
-
-import java.sql.Connection;
-
-public interface JDBCConnectionProvider {
-    public Connection getConnection();
-
-    public void closeConnection();
-}
