+++ form2-utils/src/main/java/org/bahmni/form2/exception/InvalidFormException.java
@@ -1,0 +1,8 @@
+package org.bahmni.form2.exception;
+
+public class InvalidFormException extends RuntimeException {
+
+    public InvalidFormException(String message) {
+        super(message);
+    }
+}
