+++ form2-utils/src/main/java/org/bahmni/form2/model/ControlLabel.java
@@ -1,0 +1,22 @@
+package org.bahmni.form2.model;
+
+public class ControlLabel {
+    private String value;
+    private String translationKey;
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    public String getTranslationKey() {
+        return translationKey;
+    }
+
+    public void setTranslationKey(String translationKey) {
+        this.translationKey = translationKey;
+    }
+}
