+++ form2-utils/src/main/java/org/bahmni/form2/service/Form2ReaderService.java
@@ -1,0 +1,8 @@
+package org.bahmni.form2.service;
+
+import org.bahmni.form2.model.Form2JsonMetadata;
+
+public interface Form2ReaderService {
+
+    Form2JsonMetadata read(String jsonPath);
+}
