+++ form2-utils/src/main/java/org/bahmni/form2/service/FormFieldPathService.java
@@ -1,0 +1,13 @@
+package org.bahmni.form2.service;
+
+import java.util.List;
+
+public interface FormFieldPathService {
+
+    /**
+     * @param orderedControlNames Eg: Vitals, BMI, Height as list items when Height
+     *                            control is under BMI section and form name is Vitals
+     * @return form field path Eg: Vitals.1/2-0
+     */
+    String getFormFieldPath(List<String> orderedControlNames);
+}
