+++ form2-utils/src/main/java/org/bahmni/form2/utils/ResourceNotFoundException.java
@@ -1,0 +1,9 @@
+package org.bahmni.form2.utils;
+
+import java.io.IOException;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String msg, IOException e) {
+        super(msg, e);
+    }
+}
