--- form2-utils/src/main/java/org/bahmni/module/exception/InvalidFormException.java
@@ -1,8 +1,0 @@
-package org.bahmni.module.exception;
-
-public class InvalidFormException extends RuntimeException {
-
-    public InvalidFormException(String message) {
-        super(message);
-    }
-}
