--- form2-utils/src/main/java/org/bahmni/module/model/ConceptAnswer.java
@@ -1,23 +1,0 @@
-package org.bahmni.module.model;
-
-public class ConceptAnswer {
-
-    private String displayString;
-    private String translationKey;
-
-    public String getDisplayString() {
-        return displayString;
-    }
-
-    public String getTranslationKey() {
-        return translationKey;
-    }
-
-    public void setDisplayString(String displayString) {
-        this.displayString = displayString;
-    }
-
-    public void setTranslationKey(String translationKey) {
-        this.translationKey = translationKey;
-    }
-}
