--- form2-utils/src/main/java/org/bahmni/module/model/ControlLabel.java
@@ -1,22 +1,0 @@
-package org.bahmni.module.model;
-
-public class ControlLabel {
-    private String value;
-    private String translationKey;
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    public String getTranslationKey() {
-        return translationKey;
-    }
-
-    public void setTranslationKey(String translationKey) {
-        this.translationKey = translationKey;
-    }
-}
