--- form2-utils/src/main/java/org/bahmni/module/model/ControlProperties.java
@@ -1,22 +1,0 @@
-package org.bahmni.module.model;
-
-public class ControlProperties {
-    private boolean addMore;
-    private boolean multiSelect;
-
-    public boolean isAddMore() {
-        return addMore;
-    }
-
-    public void setAddMore(boolean addMore) {
-        this.addMore = addMore;
-    }
-
-    public boolean isMultiSelect() {
-        return multiSelect;
-    }
-
-    public void setMultiSelect(boolean multiSelect) {
-        this.multiSelect = multiSelect;
-    }
-}
