--- form2-utils/src/main/java/org/bahmni/module/model/Form2JsonMetadata.java
@@ -1,24 +1,0 @@
-package org.bahmni.module.model;
-
-import java.util.List;
-
-public class Form2JsonMetadata {
-    private List<Control> controls;
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Control> getControls() {
-        return controls;
-    }
-
-    public void setControls(List<Control> controls) {
-        this.controls = controls;
-    }
-}
