--- form2-utils/src/main/java/org/bahmni/module/service/Form2ReaderService.java
@@ -1,8 +1,0 @@
-package org.bahmni.module.service;
-
-import org.bahmni.module.model.Form2JsonMetadata;
-
-public interface Form2ReaderService {
-
-    Form2JsonMetadata read(String jsonPath);
-}
