--- form2-utils/src/main/java/org/bahmni/module/service/Form2Service.java
@@ -1,15 +1,0 @@
-package org.bahmni.module.service;
-
-import java.util.Map;
-
-public interface Form2Service {
-
-    Map<String, String> getAllLatestFormPaths();
-
-    Map<String, Integer> getFormNamesWithLatestVersionNumber();
-
-    String getFormPath(String formName);
-
-    int getFormLatestVersion(String formName);
-
-}
