--- form2-utils/src/main/java/org/bahmni/module/utils/ResourceNotFoundException.java
@@ -1,9 +1,0 @@
-package org.bahmni.module.utils;
-
-import java.io.IOException;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String msg, IOException e) {
-        super(msg, e);
-    }
-}
