--- form2-utils/src/main/java/org/bahmni/module/utils/ResourceUtils.java
@@ -1,17 +1,0 @@
-package org.bahmni.module.utils;
-
-import org.apache.commons.io.IOUtils;
-import org.springframework.core.io.Resource;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-public class ResourceUtils {
-    public static String convertResourceOutputToString(Resource resource) {
-        try (InputStream inputStream = resource.getInputStream()) {
-            return IOUtils.toString(inputStream);
-        } catch (IOException e) {
-            throw new ResourceNotFoundException("Cannot load the provided resource. Unable to continue", e);
-        }
-    }
-}
